FuzzBrowse/run_20250325_053451/round_001/seed_inputs/zh-Big5-late.tentative.html_radamsa_5.html

<!doctype html>
<meta charset="utf-8">
<title><title>zh Big5 late</title><title>zh Big5 late</title><title>zh Big5 late</title><title>zh Big5 late</title></title>
<script src=/resources/testharness.js></script><title>
<script src=/resources/testharnessreport.js></title></script><script>
</script><script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "Big255", 'Expected Big5');
    }, "Check detection result");
    w.close();
    done();
};
var w = window.open("support/zh-Big5-late.sub.html");
</script>
